requirements.txt
deepface==0.0.93
Flask==3.0.3
numpy==2.1.2
opencv_python==4.10.0.84
Pillow==10.4.0
scikit_learn==1.2.2
torch==2.1.2
torchvision==0.16.2
transformers==4.37.2
